+++ Cabal/Distribution/Types/BuildInfo/Lens.hs
@@ -60,6 +60,10 @@ class HasBuildInfo a where
    ldOptions = buildInfo . ldOptions
    {-# INLINE ldOptions #-}
 
+   hsc2hsOptions :: Lens' a [String]
+   hsc2hsOptions = buildInfo . hsc2hsOptions
+   {-# INLINE hsc2hsOptions #-}
+
    pkgconfigDepends :: Lens' a [PkgconfigDependency]
    pkgconfigDepends = buildInfo . pkgconfigDepends
    {-# INLINE pkgconfigDepends #-}
@@ -227,6 +231,9 @@ instance HasBuildInfo BuildInfo where
 
     ldOptions f s = fmap (\x -> s { T.ldOptions = x }) (f (T.ldOptions s))
     {-# INLINE ldOptions #-}
+
+    hsc2hsOptions f s = fmap (\x -> s { T.hsc2hsOptions = x }) (f (T.hsc2hsOptions s))
+    {-# INLINE hsc2hsOptions #-}
 
     pkgconfigDepends f s = fmap (\x -> s { T.pkgconfigDepends = x }) (f (T.pkgconfigDepends s))
     {-# INLINE pkgconfigDepends #-}